+++ Sources/BaseFoundationiOS/Common/GlobalTask.swift
@@ -17,4 +17,12 @@ public class GlobalTask {
             }
         }
     }
+    
+    public static func performDelay(seconds: Double = 0.1) async {
+        await withCheckedContinuation { continuation in
+            Self.performDelay {
+                continuation.resume()
+            }
+        }
+    }
 }