+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `joined` array extension implemented.
 - `monthName` date extension implemented.
 - `getMonthYearName` date extension implemented.
+- `performDelay` `async` implemented.
 
 # 0.13.0